+++ History.md
@@ -9,7 +9,7 @@
 * Functions which modify roles are available both on the client and server side, but should be called on the
   client side only from inside Meteor methods.
 * `deleteRole` can now delete role even when in use, it is automatically unset from all users.
-* Functions `addRoleParent` and `removeRoleParent` were added.
+* Functions `addRolesParent` and `removeRolesParent` were added.
 * `addUsersToRoles` and `setUserRoles` now require that roles exist and will not create missing roles automatically.
 * All functions work with 1.0 arguments, but in 2.0 accept extra arguments and/or options.